[#113] Refine SaveAsImage button as a util component
