+++ frontend/src/components/utils/index.js
@@ -1,0 +1,1 @@
+export { default as SaveAsImageButton } from "./SaveAsImageButton";
